+++ src/Resources/contao/dca/tl_fernschach_spieler_konto_beitrag.php
@@ -115,6 +115,20 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler_konto_beitrag'] = array
 				'icon'                => 'bundles/contaofernschach/images/show.svg',
 				//'button_callback'     => array('tl_fernschach_spieler_konto_beitrag', 'generateShowButton')
 			),
+			'moveHaupt' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_fernschach_spieler_konto_beitrag']['moveHaupt'],
+				'href'                => 'key=moveBuchung&source=b&target=h',
+				'icon'                => 'bundles/contaofernschach/images/konto_haupt_move.png',
+				'attributes'          => 'onclick="if(!confirm(\'' . $GLOBALS['TL_LANG']['tl_fernschach_spieler_konto_beitrag']['moveHauptConfirm'] . '\'))return false;Backend.getScrollOffset()"',
+			),
+			'moveNenngeld' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_fernschach_spieler_konto_beitrag']['moveNenngeld'],
+				'href'                => 'key=moveBuchung&source=b&target=n',
+				'icon'                => 'bundles/contaofernschach/images/konto_nenngeld_move.png',
+				'attributes'          => 'onclick="if(!confirm(\'' . $GLOBALS['TL_LANG']['tl_fernschach_spieler_konto_beitrag']['moveNenngeldConfirm'] . '\'))return false;Backend.getScrollOffset()"',
+			),
 			'markiertIcon' => array
 			(
 				'label'               => &$GLOBALS['TL_LANG']['tl_fernschach_spieler_konto_beitrag']['markiertIcon'],